# chunk cache perf fixes: AddRange + batched accounting