--- a/mdata/cache/if.go
+++ b/mdata/cache/if.go
@@ -9,6 +9,7 @@ import (
 
 type Cache interface {
 	Add(metric schema.AMKey, prev uint32, itergen chunk.IterGen)
+	AddRange(metric schema.AMKey, prev uint32, itergens []chunk.IterGen)
 	CacheIfHot(metric schema.AMKey, prev uint32, itergen chunk.IterGen)
 	Stop()
 	Search(ctx context.Context, metric schema.AMKey, from, until uint32) (*CCSearchResult, error)