replaces #940

going from the old Add implementation to the new AddRange implementation
```
benchcmp pre.txt post7.txt
benchmark                   old ns/op     new ns/op     delta
BenchmarkAddAsc             1223930       1111          -99.91%
BenchmarkAddDesc1           1449968       707682        -51.19%
BenchmarkAddDesc4           1450317       708973        -51.12%
BenchmarkAddDesc64          1446361       705477        -51.22%
BenchmarkAddRangeAsc        1124          406           -63.88%
BenchmarkAddRangeDesc4      362446        196523        -45.78%
BenchmarkAddRangeDesc64     278462        157809        -43.33%

benchmark                   old allocs     new allocs     delta
BenchmarkAddAsc             5              3              -40.00%
BenchmarkAddDesc1           7              6              -14.29%
BenchmarkAddDesc4           9              8              -11.11%
BenchmarkAddDesc64          9              8              -11.11%
BenchmarkAddRangeAsc        1              0              -100.00%
BenchmarkAddRangeDesc4      2              1              -50.00%
BenchmarkAddRangeDesc64     1              0              -100.00%

benchmark                   old bytes     new bytes     delta
BenchmarkAddAsc             21684         145           -99.33%
BenchmarkAddDesc1           21720         191           -99.12%
BenchmarkAddDesc4           21729         199           -99.08%
BenchmarkAddDesc64          21701         202           -99.07%
BenchmarkAddRangeAsc        118           123           +4.24%
BenchmarkAddRangeDesc4      5503          132           -97.60%
BenchmarkAddRangeDesc64     3276          115           -96.49%
```