@Dieterbe 2018-06-14
Now also with the batched accounting, test results are in https://github.com/grafana/metrictank/pull/931#issuecomment-397260561