@Dieterbe 2018-06-18
rebased on master